refactor: split containers handler into submodules
